writeLine() - consistency

+++ Rasterizer/Rasterizer/Rasterizer.hpp
@@ -311,7 +311,7 @@ struct Rasterizer {
                 if (scale < 0.f)
                     tmp = x0, x0 = x1, x1 = tmp, tmp = y0, y0 = y1, y1 = tmp;
                 dx = x1 - x0, dy = y1 - y0, dxdy = fabsf(dx) / (fabsf(dx) + 1e-3f) * dx / dy;
-                for (sy0 = y0, sx0 = x0, iy0 = floorf(y0), delta = deltas + int(iy0 * stride); iy0 < y1; iy0 = iy1, sy0 = sy1, sx0 = sx1, delta += stride) {
+                for (sy0 = y0, sx0 = x0, iy0 = floorf(y0), delta = deltas + size_t(iy0 * stride); iy0 < y1; iy0 = iy1, sy0 = sy1, sx0 = sx1, delta += stride) {
                     iy1 = iy0 + 1.f, sy1 = y1 > iy1 ? iy1 : y1, sx1 = x0 + (sy1 - y0) * dxdy;
                     lx = sx0 < sx1 ? sx0 : sx1, ux = sx0 > sx1 ? sx0 : sx1;
                     ix0 = floorf(lx), ix1 = ix0 + 1.f;
